day29/otel-observability/services/order-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0
pydantic==2.5.2
structlog==23.2.0
